Rewrote the Stand task system to work with all stand entity action
...and added a stub for heavy attacks
...and a dash move

+++ src/main/java/com/github/standobyte/jojo/entity/damaging/projectile/ownerbound/OwnerBoundProjectileEntity.java
@@ -178,7 +178,7 @@ public abstract class OwnerBoundProjectileEntity extends ModdedProjectileEntity 
     }
     
     protected void checkRetract() {
-        if (tickCount >= (double) ticksLifespan() * retractSpeed() / (movementSpeed() + retractSpeed())) {
+        if (tickCount >= (int) ((float) ticksLifespan() * retractSpeed() / (movementSpeed() + retractSpeed()))) {
             setIsRetracting(true);
         }
     }
